Flask_integration/requirements.txt
Flask==3.1.0
minio==7.2.15
python-dotenv==1.0.0
requests==2.22.0
pdf2image==1.17.0
Pillow==7.0.0
python-docx==0.8.11
python-pptx==0.6.21
gunicorn==20.1.0
